// src/types/FormResponseRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { InputFieldResponseRequest } from './InputFieldResponseRequest.ts'

/**
 * @description Base serializer for all models.
 */
export type FormResponseRequest = {
  /**
   * @type array
   */
  response_fields: InputFieldResponseRequest[]
  /**
   * @type integer
   */
  created_by?: number | null
  /**
   * @type integer
   */
  updated_by?: number | null
  /**
   * @type integer
   */
  form_assignment: number
}