// src/types/PatchedFormResponseRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { InputFieldResponseRequest } from './InputFieldResponseRequest.ts'

/**
 * @description Base serializer for all models.
 */
export type PatchedFormResponseRequest = {
  /**
   * @type array | undefined
   */
  response_fields?: InputFieldResponseRequest[]
  /**
   * @type integer
   */
  created_by?: number | null
  /**
   * @type integer
   */
  updated_by?: number | null
  /**
   * @type integer | undefined
   */
  form_assignment?: number
}